+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Change: 'ahn restart' now does not fail if the PID file is not found
   * Change: `CallController#join` now blocks until a corresponding unjoined event is received
   * Feature: Added `Call#unjoin`
+  * Feature: `CallControll#join` can be made non-blocking by passing `:async => true`
 
 # 2.0.0.beta1 - 2012-03-07
   * Bugfix: #speak now correctly casts the argument to string if it is not SSML